solutions: ex060, float type

+++ exercises/060_floats.zig
@@ -26,7 +26,7 @@
 // operations with numeric literals, ensure the types match. Zig
 // does not perform unsafe type coercions behind your back:
 //
-//     var foo: f16 = 13.5 * 5;   // ERROR!
+//     var foo: f16 = 13.5 * 5;   // ERROR! // todo x: 类型不匹配, 无法做计算
 //     var foo: f16 = 13.5 * 5.0; // No problem, both are floats
 //
 // Please fix the two float problems with this program and
@@ -40,7 +40,7 @@ pub fn main() void {
     //
     // We'll convert this weight from tons to kilograms at a
     // conversion of 907.18kg to the ton.
-    var shuttle_weight: f16 = 907.18 * 2200;
+    var shuttle_weight: f32 = 907.18 * 2200.0; // todo x: 数字运算, 需要类型一致. 乘法结果溢出, 增大类型
 
     // By default, float values are formatted in scientific
     // notation. Try experimenting with '{d}' and '{d:.3}' to see
